requirements.txt
requests
python-dotenv
python-telegram-bot~=13.1